refactor(website): re-use mitmachen page
Some getInitialProps was kept due to them loading extra data.
We need a better way to reuse those than we currently have but it's out of scope for this.
