+++ apps/kolumna/pages/mitmachen.tsx
@@ -1,10 +1,10 @@
 import {styled} from '@mui/material'
 import {getSessionTokenProps, ssrAuthLink} from '@wepublish/utils/website'
-import {ApiV1, AuthTokenStorageKey, SubscribeContainer} from '@wepublish/website'
+import {SubscribePage} from '@wepublish/utils/website'
+import {ApiV1, AuthTokenStorageKey} from '@wepublish/website'
 import {setCookie} from 'cookies-next'
 import {NextPageContext} from 'next'
 import getConfig from 'next/config'
-import {useRouter} from 'next/router'
 import {useMemo} from 'react'
 
 import {CrowdfundingGoal} from '../src/crowdfunding/crowdfunding-goal'
@@ -26,13 +26,6 @@ type MitmachenProps = {
 }
 
 export default function Mitmachen({donate}: MitmachenProps) {
-  const {
-    query: {firstName, mail, lastName}
-  } = useRouter()
-
-  const locationOrigin = typeof window !== 'undefined' ? location.origin : ''
-  const thisLocation = typeof window !== 'undefined' ? location.href : ''
-
   const {data: subscribers} = ApiV1.useActiveSubscribersQuery()
 
   const {data: memberplans} = ApiV1.useMemberPlanListQuery({
@@ -67,19 +60,12 @@ export default function Mitmachen({donate}: MitmachenProps) {
         </GoalsWrapper>
       )}
 
-      <SubscribeContainer
-        defaults={{
-          email: mail as string | undefined,
-          firstName: firstName as string | undefined,
-          name: lastName as string | undefined
-        }}
+      <SubscribePage
         filter={memberPlans =>
           memberPlans.filter(mb =>
             donate ? mb.tags?.includes('spende') : mb.tags?.includes('crowdfunding')
           )
         }
-        successURL={`${locationOrigin}/profile/subscription`}
-        failureURL={thisLocation}
         fields={['firstName']}
         termsOfServiceUrl="/agb"
         donate={() => !!donate}
